Move modelAPI and modelStorage to a single package 'model'
